backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
Flask
flask-cors
